requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlalchemy==2.0.43
asyncpg==0.30.0
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.3.0
httpx==0.28.1
python-jose[cryptography]==3.5.0
pytest==8.4.2
pytest-asyncio==1.2.0
